Refactored models package to break it up into more logical packages
