--- models/doc.go
@@ -1,3 +1,0 @@
-// Package models stores various models for the collection service
-// Todo: maybe split this package up if it makes more sense
-package models
