+++ output/doc.go
@@ -1,0 +1,2 @@
+// Output contains models for the output data from the collection service
+package output
