+++ realtime/doc.go
@@ -1,0 +1,2 @@
+// Realtime stores various models for the results returned by the GTFS realtime service
+package realtime
