+++ static/doc.go
@@ -1,0 +1,2 @@
+// Static contains models and access methods for static data in the delay database
+package static
